+++ src/mcmod.info
@@ -2,9 +2,7 @@
   "modid": "vshop",
   "name": "VillagerShops",
   "description": "Set up Mobs of any type as admin shop. Shops will use a Inventory-Menu to buy/sell items",
-  "version": "0.1",
-  "mcversion": "1.10.2",
+  "version": "0.2",
   "url": "https://github.com/DosMike/VillagerShops",
-  "authorList": [ "DosMike" ],
-  "useDependencyInformation": false
+  "authorList": [ "DosMike" ]
 }]
